+++ packages/lumx-core/src/js/utils/index.ts
@@ -1,2 +1,3 @@
 export * as classNames from './classNames';
 export * from './events';
+export * from './selectors';